requirements.txt
Flask==2.2.3
requests==2.28.2
pyyaml==6.0